-Updated POCO from 1.9.4 to 1.10.1 -Added Configuration file support -Fixed a potential overflow Issue thanks to JonnyBravo -Removed deprecated code from AlefPacket -Initial DB implementation. More information about the specifics will be given at a later time, it is still in it's infancy, but I think I know the direction I will be taking it. However, it is subject to change. -Misc other cleanups.

+++ AlefLogin/AlefLoginClientConnection.h
@@ -119,7 +119,7 @@ private:
 		memmove(packetData, packetData + 7, packet->getSize());
 
 		for (int i = 0; i < (packet->getSize() >> 3); ++i)
-			blowfish_decrypt((isServerPacket) ? cryptSession->serverCtx : cryptSession->clientCtx, (UInt32*)(packetData + i * 8), (UInt32*)(packetData + i * 8 + 4));
+			blowfish_decrypt((isServerPacket) ? cryptSession->serverCtx : cryptSession->clientCtx, (UInt32*)(packetData + i * (__int64)8), (UInt32*)(packetData + i * (__int64)8 + (__int64)4));
 
 		if (packetData[0] != 0xD6)
 			return false;
